requirements.txt
djangorestframework==3.13.1
Django==4.0
django-extensions==3.1.5